requirements.txt
asgiref==3.2.10
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.1
django-autocomplete-light==3.7.0.dev0
django-crispy-forms==1.9.2
django-filter==2.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.1
drf-yasg==1.17.1
idna==2.10
inflection==0.5.1
irekua-database==0.1.2
irekua-dev-settings==0.1
itypes==1.2.0
Jinja2==3.0.0a1
jsonschema==3.2.0
MarkupSafe==2.0.0a1
numpy==1.19.2
packaging==20.4
Pillow==7.2.0
psycopg2-binary==2.8.6
pycountry==20.7.3
pyparsing==3.0.0a2
pyrsistent==0.17.3
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
selia-templates==0.1.2
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
timezonefinder==4.4.1
uritemplate==3.0.1
urllib3==1.25.10

